+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -70,7 +70,14 @@ Improvements to clang-tidy
   Warns if global, non-trivial objects with static storage are constructed, unless the 
   object is statically initialized with a ``constexpr`` constructor or has no 
   explicit constructor.
+  
+- New `fuchsia-trailing-return
+  <http://clang.llvm.org/extra/clang-tidy/checks/fuchsia-trailing-return.html>`_ check
 
+  Functions that have trailing returns are disallowed, except for those 
+  using decltype specifiers and lambda with otherwise unutterable 
+  return types.
+    
 - New alias `hicpp-avoid-goto
   <http://clang.llvm.org/extra/clang-tidy/checks/hicpp-avoid-goto.html>`_ to 
   `cppcoreguidelines-avoid-goto <http://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines-avoid-goto.html>`_